refactor(ui): extract work graph queue component
