+++ frontend/src/format.ts
@@ -1,0 +1,19 @@
+export function formatTime(value: string): string {
+  if (!value) {
+    return "unknown";
+  }
+  const date = new Date(value);
+  if (Number.isNaN(date.getTime())) {
+    return value;
+  }
+  return new Intl.DateTimeFormat(undefined, {
+    month: "short",
+    day: "2-digit",
+    hour: "2-digit",
+    minute: "2-digit",
+  }).format(date);
+}
+
+export function labelForStatus(status: string): string {
+  return status.replace("_", " ") || "unknown";
+}
